# backend/openapi.yml
...
openapi: 3.1.0
info:
  title: FocusMe Cloud API
  version: 1.0.0
  description: |
    Cloud backend API for FocusMe — provides user authentication,
    cross-device plan synchronization, and family dashboard features.

    ## Authentication

    Most endpoints require a Bearer JWT access token in the
    `Authorization` header. Obtain tokens via `/auth/login` or
    `/auth/register`. Access tokens expire after 15 minutes;
    use `/auth/refresh` to rotate.

    ## Rate Limiting

    - Auth routes: **10 requests/minute** per IP
    - All other routes: **100 requests/minute** per IP
    - Exceeded → `429 Too Many Requests` with `Retry-After` header
  contact:
    name: FocusMe Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.focusme.app/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Development

# ── Security Schemes ──────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token obtained from /auth/login or /auth/register.
        Include as `Authorization: Bearer <token>`.

  # ── Reusable Schemas ──────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code constant for programmatic handling
          example: "AUTH_INVALID_CREDENTIALS"
          enum:
            - AUTH_INVALID_CREDENTIALS
            - AUTH_EMAIL_EXISTS
            - AUTH_TOKEN_EXPIRED
            - AUTH_TOKEN_INVALID
            - AUTH_REFRESH_REUSE
            - SYNC_VERSION_CONFLICT
            - SYNC_PLAN_NOT_FOUND
            - FAMILY_INVITE_EXPIRED
            - FAMILY_ALREADY_MEMBER
            - VALIDATION_FAILED
            - RATE_LIMIT_EXCEEDED
            - FORBIDDEN
            - NOT_FOUND
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error message (do not match on this — use code instead)
          example: "Authentication failed: invalid credentials"
        details:
          type: object
          description: Optional structured error context (e.g., field-level validation errors)

    # ── Health ──────────────────────────────────────────────
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "SecureP@ss123"
        display_name:
          type: string
          example: "Jane Doe"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "SecureP@ss123"
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: The refresh token from a previous login or refresh
          example: "eyJhbGciOiJIUzI1NiJ9..."
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: JWT access token (15 min TTL)
          example: "eyJhbGciOiJIUzI1NiJ9..."
        refresh_token:
          type: string
          description: JWT refresh token (30 day TTL)
          example: "eyJhbGciOiJIUzI1NiJ9..."
        token_type:
          type: string
          enum: [Bearer]
          example: "Bearer"
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 900
    AuthResponse:
      type: object
      required: [user, access_token, refresh_token, token_type, expires_in]
      properties:
        user:
          $ref: "#/components/schemas/User"
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
    User:
      type: object
      required: [id, email, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
        display_name:
          type: string
        created_at:
          type: string
          format: date-time

    # ── Plans ───────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      summary: Create a new user account
      description: |
        Registers a new user with email and password. Password is hashed
        with Argon2id (m=64MiB, t=3, p=4). Returns user object and JWT
        token pair.
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error (invalid email, password too short)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Validation error: password must be at least 8 characters"
                code: 400
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Conflict: email already registered"
                code: 409
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              description: Seconds until the next request is allowed
              schema:
                type: integer
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      operationId: loginUser
      summary: Authenticate and receive JWT token pair
      description: |
        Validates email/password against stored Argon2id hash. Returns
        access token (15 min) and refresh token (30 days).
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Authentication failed: invalid credentials"
                code: 401
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Rotate refresh token and get new token pair
      description: |
        Validates the refresh token, revokes it, and issues a new
        access/refresh pair. If a revoked token is resubmitted (reuse
        detection), ALL refresh tokens for the user are revoked, forcing
        re-authentication on all devices.
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token pair refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Invalid or revoked refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Authentication failed: refresh token revoked — possible token theft"
                code: 401
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          headers:
            Retry-After:
              schema:
                type: integer
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Plans ────────────────────────────────────────────────────
...